MA-10527-1 [Android] add option first_partition_offset which means the offset of first partition
BSP have change the env offset to 2M which for TEE. We plan to reserve 8M for bootloader.
The offset of 'boot' partition is 2M for we set disk_alignment as 2M.
If we set disk_alignment to 8M. we have about 5 partition which have 1-2M size.
So It may waste 30+M space if we set alignment as 8M..

I add a option which represent the offset of first partition and we still use 2M
as alignment size.

Change-Id: Id1cab77d1a3d59564f2d14946cdcf9da4fe00b07
Signed-off-by: zhang sanshan <sanshan.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/3135
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>