MA-12969-1 Add -D option for flash scripts to specify image directory
for fastboot_imx_flashall.sh and fastboot_imx_flashall.bat, -D option is
added to specify the directory of images, so only one copy of script is
needed on pc, and a directory of NFS can be specified.

also, add code to handle the situation that a platform doesn't support
dual slot but a slot is specified in command.

for fastboot_imx_flashall.bat, make help infromation the same when
executed in powershell and cmd.

Change-Id: I1031dbb92120ec1a67c31326a87260a7704ba230
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>