MA-12969-2 Fix errors when no -D option provided, add -s option.
for last commit of fastboot_imx_flashall scripts, -D option was added,
but if -D option not provied when use these scripts, errors will be
encountered. this commit fix this error.

Also, -s option is added, it can be used to specify the serial number of
board connected to computer, this helps when more than one board is
connected to  computer.

For fastboot_imx_flashall.bat, handle the situation when illegal
options are provided.

Change-Id: I448a3fde314cd7b7d4cc413f474d2bdd3bf935e1
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/6292
Reviewed-by: zhang bo <bo.zhang@nxp.com>