MA-13537 build u-boot image for uuu to flash images
uboot image used by uuu should be different from the one flashed into
target boot device. prebuild uboot image for uuu are provided.

For the convenience of adding or removing features on uboot for uuu,
the build system need to be able to build out the uboot image for uuu
in response to the modifications in code.

for imx8, defconfig file based on Android team files are used, because
ab slot feature need to be used.

Change-Id: I158148fe698619a47e448bd175adef394b81f715
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>