MA-13603-1 lock the board after use uuu related script to flash boards
for the convenience of development, uuu related scripts were provided
with setting lock status is not the default situation.

now for product consideration, set lock status as default with the
modification of this patch.

also, uuu version is changed in scripts, for 1.1.81 start to support
seperate shell command, it is a feature used in uuu_imx_android_flash
scripts.

in uuu lst scripts, the uboot which is loaded into ram by uuu is now
changed to be different from the one flahsed into board, and it doesn't
support lock/unlock command, doesn't have hardcoded m4 partition for
imx7ulp.

Change-Id: I95a5aa016c64e3bb8a337b1ef53dd712464d1a86
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>