MA-13940-1 copy spl and uboot separately and set rbidx for uboot
To enable dual bootloader, containers respectively contain SPL and
uboot proper are not composed together. In this patch, the container
contains uboot is  copied to $OUT directory separately from the one
contains SPL.
Also, a makefile variable is defined to specify the rollback index
value for the container contains u-boot proper.

Change-Id: I4231bca80d3f1bb3393f9266f1bd4ba2072a243d
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/7140
Reviewed-by: ji.luo <ji.luo@nxp.com>