MA-13940-3 add support for dual bootloader in uuu flash scripts
In previous development, some features is not developed for or8.1, so
some patches on uuu flash script is not imported to imx_or8.1 branch with
cherry-pick. in fact, uuu scripts in imx_pi9.0 branch is a superset of
the those in imx_or8.1 branch with their function.

This patch just copy the uuu flash scripts from imx_pi9.0 branch to
imx_or8.1 branch, in which flashing dual bootloader is supported.

Change-Id: Ifc3aeeb8bcb82084d2f92e0793cc31d570766c70
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/7179
Reviewed-by: ji.luo <ji.luo@nxp.com>