MA-13979 align flash script behaviour for cmd and powershell on windows
On windows, if a subroutine encouters an error, the subroutine will
exit, but the script will keep on executing to the end, so even if some
images are not flashed, the script will output info prompt that all
images are successfully flashed. The code call subroutines need to check
whether error is ecountered in subroutine, and if there is, just exit.
Since "exit" behaves different for cmd and power shell, the code can only
exit one levle by one level.

On linux, when flash script can successfully flash all images, there is
some error log. it's because sometimes shell variables used in condition
judging statement may have empty value, this kind of variable need to be
quoted.

Change-Id: I22daac682611b04f0e19baabf872d599e3635333
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/7233
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>