MA-14010 change the target image name for evk_8mm
Before this change, two uboot image names for evk_8mm are as below:
    u-boot-imx8mm-ddr4.imx    for boards with ddr4 on them
    u-boot-imx8mm.imx         for boards with lpddr4 on them
Four dtbo and vbmeta image names are as below, first three is for boards
with lpddr4 on them, the last one is for boards with ddr4 on it.
    dtbo-imx8mm-lpddr4.img ------- vbmeta-imx8mm-lpddr4.img
    dtbo-imx8mm-m4.img ----------- vbmeta-imx8mm-m4.img
    dtbo-imx8mm-mipi-panel.img --- vbmeta-imx8mm-mipi-panel.img

    dtbo-imx8mm.img -------------- vbmeta-imx8mm.img

This patch will change the name of dtbo and vbmeta image.
dtbo-imx8mm-lpddr4.img change to be dtbo-imx8mm.img
dtbo-imx8mm.img change to be dtbo-imx8mm-ddr4.img

So for boards with lpddr4, no "ddr" related info in it.
images used by boards with ddr4 on it have "ddr4" in image name.

Making this change can avoid modification in image flash script and
release doc.

Change-Id: I9a7c024be3173ee837b451524ad6e5aacfd64116
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/7275
Reviewed-by: Tian Yang <yang.tian@nxp.com>