MA-14011 add uuu SDPV support for images enabled SPL feature
in previous version, invoke uuu with an offset to specify the part to
be loaded by SPL. It is corresponding to SDPU command.

Now for images enabled SPL, uuu is invoked with a parameter to specify
that SPL should be skipped. It is corresponding to SPDV.

Since we may flash images release before, so SDPU and SDPV both need to
be supported.

Change-Id: I106f100dd2d2b0d2cda28031ed03b81899706943
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/7303
Reviewed-by: ji.luo <ji.luo@nxp.com>