MA-14014 fix the issue that xen uboot will be deleted during buid time
when dual bootloader is not enabled, u-boot-$$UBOOT_PLATFORM.imx will be
copy to $OUT directory. for imx6 and imx7, it the final uboot image. for
imx8, there is post process to to ingetrate other firmware into the
final uboot, and u-boot-$$UBOOT_PLATFORM.imx in $OUT directory will be
overriden.

After dual bootloader is enabled, u-boot-$$UBOOT_PLATFORM.imx will not
be overriden in post process, because spl-$$UBOOT_PLATFORM.bin and
bootloader-$$UBOOT_PLATFORM.img will be copied to $OUT directory.

previous patch didn't handle the condition of xen uboot, xen uboot
doesn't have post process but is deleted as it has. it need to be fixed.

Change-Id: If45da02ea1ff6d906cf707dff6ae6cdfa9a5cb5d
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/7284
Reviewed-by: Wang Haoran <elven.wang@nxp.com>
Reviewed-by: ji.luo <ji.luo@nxp.com>