MA-14119 dump binary partition table as judgement rule for some features
In previous commit, the name of files in the directory which contains
the image files to be flashed to judge whether dual bootloader, recovery,
dtbo and dual slot are supported or not. If image files for different
platforms are put together in one directory, this would cause judgetment
error.

This patch dump the content of binary file and find specific strings as
judgement principle.

Also, erase userdata partition is moved to the second-to-last line in
the generated uuu script. Developers can force to stop the flashing
process when the script is executed to the erase userdata stage since
we may just want userdata be corrupted but not fully erased in the
process of flashing images

Change-Id: I1665db2a6755c1dbc656bfb525d8c612b5152753
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/7406
Reviewed-by: Wang Haoran <elven.wang@nxp.com>