MA-14688 flash bootloader first for legacy device to overwritten GPT
when a sdcard has i.MX8 images in it, both GPT and backup GPT are in it.
To use this card to flash images for legacy devices, only MBR and backup
GPT. previous GPT will be loaded and used. This will cause the flash
process terminate with an error.

flash bootloader first, previous PTEs will be overwritten and this error
is avoided.

Change-Id: Ie0fa6d132e2a7b71a7f65e784328cdb61feda932
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/8364
Reviewed-by: ji.luo <ji.luo@nxp.com>