MA-10879 Add lcd_density in bootargs to over write default value
All imx8 platform connect to HDMI as default display.
When it is connected to MIPI display, the density is different.
Add lcd_density in bootargs to control display density.

Need add default "android.lcd_density" in bootargs, or
ro.sf.lcd_density will be empty.

This patch is used to replace below patch in system/core:
82abe411f MA-10879 Add lcd_density in bootargs to over write default value.

Change-Id: I35141afe492e1829a6ba7440bd48fe61161aeaee
Signed-off-by: Jindong <jindong.yue@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/9554
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>