MA-15355 fix selinux error in Q preview
In init Pi9.0, init doesn't try to relabel a link file with the type of
"block_device" to the type "dtbo_block_device". Also. The uevent file
vold access is with the type of "sysfs"

In Q preview, init tries to relabel a link file with the type of
"block_device" to the type of "dtbo_block_device", and the uevent file
vold access is with the type of "sysfs_block_devices".

To adapt to these changes, two allow rules are added to eliminate
related selinux errors.

Change-Id: I9de3ace68d419e3021648b2226f6e8372578314f
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/9847
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>