MA-15339-4 trigger the request on control of display when using GSI
when GSI is used, "vendor.all.system_server.ready" will not be set to
trigger that the A core request control of display.

set "vendor.all.system_server.ready" to 0 zero at a very early stage,
if A core boot completed and this property is still 0, then trigger the
request on control of display.

Change-Id: I884668d46f7105bcda6b2bed9cbda1127975243d
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/10192
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>