MA-15885-1 check the legality of the parameter after -u and -d
add logic to judge the legality of the parameter after -u and -d

also, for uuu_imx_android_flash.sh and fastboot_imx_flashall.sh
the definition of functions is moved to the front part, the function
definition will not be executed.
for batch scripts, the so-called functions must be defined follow the
process flow which will invoke the functions, or the funtion definition
will be executed.

add arrays to collect the features supported. only 7ulp, 8mm, 8mn, 8mq,
8qm, 8qxp are listed. other soc related board are removed from the help
info, but keep the related logic in process code in case they are used
one day.

batch scripts encoded in ANSI instead of UTF-8, so the table in help
message can be output correctly.

Change-Id: I0bb321fc533097e8eec63d5f03708f30075a7e8b
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/10581
Reviewed-by: zhang bo <bo.zhang@nxp.com>
Reviewed-by: ji.luo <ji.luo@nxp.com>