MA-15722 build uboot and kernel independently from Android rootfs
Configurations for uboot/kernel is decoupled from the ones for Android rootfs
uboot/kernel dedicated configurations are put in below two files:
    UbootKernelCommonConfig.mk
    UbootKernelBoardConfig.mk
Shared configurations are put in below file:
    SharedBoardConfig.mk

Some deprecated configuration items are removed. they are listed as below, the
first two are deprecated in the AOSP core makefiles, so related configuration
need to be removed. for the other three, it's only assigned value in device/fsl
git, not used elsewhere by searching in the code.
    ARCH_ARM_HAVE_TLS_REGISTER
    TARGET_CPU_SMP
    TARGET_HAVE_VIV_HWCOMPOSER
    TARGET_SELECT_KEY
    IMX_CAMERA_HAL_V3

two compilation FLAG is removed, which were used to eliminate
compilation error when builkd GPU driver, now the GPU driver is refined.

in AndroidUboot.mk, "==" is replaced as "=" to do equal test, `` is used
to replace $(shell) for compatiblity.

dtb compilation is moved from dtbo.mk to kernel.mk. uboot.mk and
kernel.mk are disconnected with the AOSP core makefiles.

a new Makefile and imx-make.sh is added to build uboot/kernel.

Change-Id: I792f9671195425d1a9cfba52033082cca8d06c60
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/10416
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: Jindong Yue <jindong.yue@nxp.com>
Reviewed-by: Wang Haoran <elven.wang@nxp.com>