MA-14514-2 initialize batch variable at the begining
run batch script with "cmd", the variables in the batch file will be
saved by "cmd". so the value set in the batch script may affect the
execution of this script in next time.

"dryrun" is set to one if "-dryrun" option is provided, when run this
script again, even "-dryrun" not provided, "dryrun" variable still has
the value of 1. Need to initialize the "dryrun" as 0 at the begining.

powershell does not have such issue.

Change-Id: If3e4cedb3400374b785ba0b9bdd14990cfb98919
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/10744
Reviewed-by: ji.luo <ji.luo@nxp.com>