MA-16078-2 Support non product partition build
Add variable 'IMX_NO_PRODUCT_PARTITION' to support non
product partition build. This commit will remove the
product partition in GPT and enlarge the system partition
to 2816MB as it may get larger in the further.

This config is to align the image overlay in Android9, but
please note the partition size of system and vendor are
different. This config will be disabled by default.

Test: build and boot on imx8qxp and imx8mm.

Change-Id: Iac02a5b8a54de54d0b734ed1b2bdf783430daf7c
Signed-off-by: Ji Luo <ji.luo@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/10799
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>