MA-16457-4 change super partition to 7GB
5GB is not enouth for images with GSI, chane it to 7GB.

When dynamic partition enabled, there is no physical partition anyway,
so there is no need to keep the files for the condition when dynamic
partition is enabled and there is no product image.

an item in fstab file is useless since recovery mode can't access this
file, remove this item.

Change-Id: I2566458f38592de366a98a14458c5b9255e3bd44
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/11480
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>
Reviewed-by: Wang Haoran <elven.wang@nxp.com>