MA-16668-1 Fix kernel hang after enabling debug configs
The Image becomes bigger after enabling some debug options,
ajust the ramdisk offset to reserve more space for kernel
and fdt. For imx8m, the ramdisk will have 64MB offset to
the kernel. For imx8q, we put the ramdisk after 'encoder_boot'
reserved memory as there is a memory hole can be used to
hold the ramdisk.

This CL also enlarge the boot partition to 64MB in gpt or build
will fail after enabling kasan for imx8q.

Test: boots with 'kasan' enabled.

Change-Id: Ie8188814862f830fae67e6dcb608087e202f196f
Signed-off-by: Ji Luo <ji.luo@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/11679
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: Wang Haoran <elven.wang@nxp.com>
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>