MA-16457-9 demonstration of retrofit dynamic partition
This is a demonstration of OTA update from images with which dynamic
partition feature is not supported to images which support dynamic
partition feature.

Dynamic partition metadata are put in system partition.
BOARD_SUPER_PARTITION_SIZE equals the sum of the partitions which will
be logical partition after OTA update, reserve some storage for
metadata, OARD_NXP_DYNAMIC_PARTITIONS_SIZE is the value of the ramaining
storage size.

This is a demo of update from android10.0.0_1.0.0 to android10.0.0_2.0.0
or higher. The uboot image for Pi9.0 release and before does not load
ramdisk to boot up android, instead, it will pass bootargs to have
kernel mount the system partition, but the system partition will be
logical partition and kernel can not recognize it.

Change-Id: I85b0f70d509e22cfa8ac756d62abe8171cad9bbb
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/11709
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: Wang Haoran <elven.wang@nxp.com>