MA-16457-10 do not pass the partition size for lpmake
simglen was used to get the raw image size of system.img, vendor.img and
product.img, this size will be set to the partition size in lpmake
command.

lpmake now can itself detect the size of the partition if a image is
specified for the partition, lpmake will use similar logic as in simglen
to get the size.

Also, simglen will not be copied to ${OUT} directory.

Change-Id: I433ed2a545b78829ecd3a754cc5d8368a066d27e
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/11729
Reviewed-by: Wang Haoran <elven.wang@nxp.com>