MA-16694 exit if any errors encountered
use "-e" option for the linux shell script, so that any errors
encountered will cause the execution termination. the function
"whether_in_array" now can't return 1, since this will end the execution
of the script.

fix the set active issue in fastboot_imx_flashall script.

remove the limitation of image directory provided for
fsl-sdcard-partition.sh must ended with a slash.

also, an option of "-super" is provided, if this option is used, the
script will not generate super.img, it will directly use the super.img
in the directory together with other images.

Change-Id: If00120fe9dca92eac14408f5f3f279aebc293beb
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/11775
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: ji.luo <ji.luo@nxp.com>