MA-16793 correct the path for spi norflash and nand in sysfs
The path for spi norflash and nand in sysfs are not correct in
genfs_contexts with kernel v5.4.3. so the related files are labeled
with "sysfs", not "sysfs_block".

Below case of "android.security.cts.SELinuxHostTest#testNoBugreportDenials"
in the module of"CtsSecurityHostTestCases" will have dumpstate read the "stat"
file of a block device in sysfs to get the statistics of the block
device, and then check whether the operations conducted by dumpstate are
denied by selinux.

dumpstate is not allowed to read files labled with "sysfs", so the files
are labeled as "sysfs_block" here and allow dumpstate to read them.

Change-Id: I42613a4116517d87a4a4603e357898695dcb4c82
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/11796
Reviewed-by: ji.luo <ji.luo@nxp.com>