MA-16457-11 fix errors when enable dynamic partition feature
Specify the vendor partition size only when dynamic partition feature is
not enabled.
There is no need to specify the vendor partition size when dynamic partition
feature is enabled, or the vendor image will have a limitation at built time,
although its size is still dynamic when doing OTA. Also, with a size for vendor
partition, there is free space for it, so when doing remount, overlayfs is not
used for vendor partition.

set "BOARD_BUILD_SYSTEM_ROOT_IMAGE" to be true when dynamic partition
feature is not enabled.
this variable should not be set to true when dynamic partition is
enabled. Although now system image will integrate root image by default,
this variable should be set to make the vbmeta image have commandline
descriptor in it which is used to setup dm-verity for system partition.
Without this variable, Android automotive will boot up that system
partition is not checked with dm-verity.

In this patch, the common info about product partition is moved to
common place.

Change-Id: Ie29c9b8946a72d2ac6d58832a7394d4bcdd643c5
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/12125
Reviewed-by: Wang Haoran <elven.wang@nxp.com>