MA-17355 remove temporary files generated by lpmake.exe
when generate super.img with lpmake, the images in logical partitions
like system.img will firstly be converted to raw image format from
sparse image format and written as temporary files, data in these
temporary files are then written to super.img.

On windows, the temporary files are stored in %TMP% directory and will
not be removed automotically, this patch remove the temporary files
after super.img is generated.

Change-Id: I7fecbde1b3d47a473ed8a56a410e62fafe954594
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/12549
Reviewed-by: ji.luo <ji.luo@nxp.com>