MA-17361 support to flash xen images in fastboot script
support to flash xen images in fastboot_imx_flashall script. "xen" uboot
feature is added. When flash xen images, do not flash bootloader0 images
and do not reboot after uboot images are flashed.

for uuu_imx_android_flash script, if "xen" is used as dtb feature, it
will also be used for uboot, and "xen" is not allowed to be used
together with "-u" option, allow "-u xen" to be used in this patch.

Change-Id: I7edbe9d16af7cde8d50fee070e317f3017236809
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/12560
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/12609
Reviewed-on: http://androidsource.ap.freescale.net/project/12788