MA-17551 [Android] metadata partition will been fail when mount in android11
although we define “formattable” in fstab or dts for metadata, however fs_mgr do not make it as ext4 format when boot up
If the first 4K is “0” or “ff”, it will been think as wiped. it can been format as ext4 if set as wiped. erase metadata here to fix this issue.

Change-Id: Idb8db9c18e9fdf9584f17af3777ff91aa5ba15bd
Signed-off-by: zhang sanshan <pete.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/12945
Reviewed-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/13318
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>