MA-17600 fix image flash issue when images are in a path containing spaces
In uuu_imx_android_flash.bat, %cd% is used to get current working
directory, and it's value is used together with the image names deduced
from the parameters used when execute this batch file to get the
absolute path of the images to be flashed and create link files based on
this absolute path.

If the path of current working directory contains spaces, the operation
of create link files will not work as expected, add quotation marks for
the value of %cd% to avoid this issue.

Remove the description of "-super" option which is now not supported.

Change-Id: I0ad1bfb7a2fa3747fb23991d94f350679e8f7e60
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/13157
Reviewed-by: ji.luo <ji.luo@nxp.com>
Reviewed-by: Wang Haoran <elven.wang@nxp.com>