MA-17859 fix evk_7ulp OTA package build and apply issue
rename "BOARD_PREBUILT_DTBOIMAGE" to be "TEMP_BOARD_PREBUILT_DTBOIMAGE"
in imx-recovery.mk, as this "BOARD_PREBUILT_DTBOIMAGE" is used to
specify the dtbo image in target package file, which is then used to
generate OTA package. it should not be changed when build the target
files.

add the partitions which are updated in OTA procedure in the fstab file
because "ota_from_target_files" reads the partition types from this
file. recovery partition is also listed in it, then the script to
update recovery partition can be successfully generated.

"BOARD_PREBUILT_DTBOIMAGE" specified a dtbo image with hash footer, and
when generate the recovery image in target package file, the recovery
image does not have a valid dtbo image integrated. use
"BOARD_PREBUILT_RECOVERY_DTBOIMAGE" to specify on dtbo file without hash
footer.

define the "IncrementalOTA_InstallEnd" function to directly add dtbo and
vbmeta image to the incremental OTA package.

Change-Id: I960308e76e3a6932d5a087ed360bd1384e804ffc
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/13496
Reviewed-by: Wang Haoran <elven.wang@nxp.com>