MA-17661 update super image when build ota and target file package
Set superimage_and empty superimage to be the pre-requisites of
otapackage and target-files-package.
Set imx_vbmetaimage to be the pre-quisite of target-files-package.

When OTA package and target files package are built according the
commands provided in user guide, images in ${OUT} directory are in a
good state of avb and dm-verity, so after the images are flashed, the
system can boot up.

Change-Id: I1224273630ae137da35a78e9c7a17906f49f9af9
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/13505
Reviewed-by: Wang Haoran <elven.wang@nxp.com>