MA-17872-1 remove usb controller info from udisk fstab entry
udisk is mounted by vold on Android, vold monitors the uevent of block
type, if the block uevent path can be matched with the one provided in
fstab as mount device, then the block device can be mounted.

remove the controller related info in the fstab entry, remaining part is
enough to do the pattern match, and the sysfs path for controller is a
bit mutable.

Change-Id: Ie0c5db3607a707b86b7b1ef522d4232437a59b03
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/13589
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>