MA-17448[Android] Disable cpuidle when there are INTERACTION power hint and enable after some time
CpuIdle affect the janky frames because it make the response of interrupt delayed.
Here disable CpuIdle by write 0 to /dev/cpu_dma_latency when get input event(INTERACTION) and
turn on again after some time. So it will not affect the power when when there is
no touch input event.

Change-Id: Ie17211a8819b04ee1c5c823fb95817779e7f62c2
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/13624
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>