MA-18028 support to flash sdcard connected to host via sdcard slot
When sdcard is connected to the host via the sdcard slot on the host PC,
this sdcard is detected as mmc block device. The partition device name
can be generated with this device name followed with a number of
partition index.

When sdcard is connected to the host via the USB adapter, it is
enumerated as scsi disk device. The partition device name can be
generated with this device name followed with a character 'p' and then
the partition index.

In this commit, the partition device names are made more common. based
on the name of the device provided in command line to be flashed, find
the file which has the same major device number but with the minor
number greater by 1, remove the the '1' behind this file name to get the
base name of partition devices.

Change-Id: I0402a12099235995d58405fe838cac3616acec49
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/13765
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: Wang Haoran <elven.wang@nxp.com>