MA-18168 avoid to make default target in kernel.mk
the command in imx-make.sh used to build the target ${build_vvcam}
and ${build_galcore} can be executed even if the two target is not
set, in this condition, the first target in kernel.mk as the default
target is then buit, this may lead to and error when we only want to
build bootloader. Check whether kernel modules need to be built before
execute this command.

Also fix a typo in this commit.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: I83e7924e73b1442e5773bd1f28110de47833e637
Reviewed-on: http://androidsource.ap.freescale.net/project/13897
Reviewed-by: Wang Haoran <elven.wang@nxp.com>