MA-18028 support to flash sdcard connected to host via sdcard slot
When sdcard is connected to the host via the sdcard slot on the host PC,
this sdcard is detected as mmc block device. The partition device name
can be generated with this device name followed with a number of
partition index.

When sdcard is connected to the host via the USB adapter, it is
enumerated as scsi disk device. The partition device name can be
generated with this device name followed with a character 'p' and then
the partition index.

In this commit, the partition device names are made more common. based
on the name of the device provided in command line to be flashed, find
the file which has the same major device number but with the minor
number greater by 1, remove the the '1' behind this file name to get the
base name of partition devices.

Change-Id: Ic4cc896c8a8ae4889e8f08df72cdefb9a9c95f33
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/14058
Reviewed-by: Wang Haoran <elven.wang@nxp.com>