MA-18028-1 fix potential issues caused by delimeter
the delimeter is changed to '\n' for temporary usage, after that, it
should be restored in every execution path so in further execution flow
it can be well used.

add a "nocreat" symbol to the dd command so it does not try to create
the output file if it does not exist.

Change-Id: Icf292dbbf0d55838377d83ea8d2cba64e07411c1
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/14089
Reviewed-by: Wang Haoran <elven.wang@nxp.com>