MA-17820-4 allow kernel to start usermode helper binary
Below vts cases test on kernel key management, usermode helper binary is
started to try to finish the construction of keys in the test.

The usermode helper binary is set in another patch, this commit add
selinux rules to allow kernel to start this binary so below cases can
pass.

  module: VtsKernelLtp
  case: VtsKernelLtp#cve.cve-2017-12192_64bit
        VtsKernelLtp#cve.cve-2017-15299_64bit
        VtsKernelLtp#cve.cve-2017-15951_64bit
        VtsKernelLtp#syscalls.keyctl07_64bit
        VtsKernelLtp#syscalls.request_key03_64bit
        VtsKernelLtp#dio.dio14_32bit

Change-Id: I1410d7c352ba414d58d9b6d23dbdc8f23b3066e4
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/14197
Reviewed-by: Wang Haoran <elven.wang@nxp.com>