MA-18392 enable OTA postinstall function
Define postinstall command for vendor partition, the command is
imx_ota_postinstall, it is now used to update the image in the
partition named bootloader0.

Copy the corresponding image of bootloader0 partition to vendor
partition if "IMX_OTA_POSTINSTALL" is set.

set selinux context for properties and files which are accessed
by postinstall command.

Change-Id: Icb33a8a1e4a2bd3cebf42e9b1159bc7934f79c90
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/14440
Reviewed-by: Wang Haoran <elven.wang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/14456
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>