MA-18429-1 use variables instead of hardcoded path strings
To make the configuration code more common, use variables to replace
hardcoded paths.

For variable "IMX_DEVICE_PATH", assign it a value based on the value of
"$(MAKEFILE_LIST)" instead of directly hardcode the path string of
current file. Use it in propriate places to copy configuration files.

Define a variable "CONFIG_REPO_PATH" with value "device/nxp", it is the
directory of this repository, then use this variable when need this path
string.

Use the variable "PRODUCT_DEVICE" instead of hardcode the output string
path. In the device makefile, $(firstword $(PRODUCT_DEVICE)) is used
because inherit-product function invoked in the same file append other
stings to the value of this variable.

"BOARD_TYPE" is defined by not used, remove it.

evk_8mm_drm.mk and evk_8mq_drm.mk is not modified in this commit.

Change-Id: I4894c5bd657390c0423eabf04198f90848fb8e07
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/14550
Reviewed-by: Wang Haoran <elven.wang@nxp.com>